Address any regressions caused by the structural changes I made to facilitate NR state changes and testing. Tested end to end.
