+++ api/namex/services/name_request/abstract_name_request.py
@@ -3,9 +3,11 @@ from pytz import timezone
 
 from namex.constants import request_type_mapping
 
-from namex.models import db, NRNumber, User
+from namex.models import db
+from namex.models.nr_number import NRNumber
+from namex.models.user import User
 
-from namex.services.name_request.utils import handle_exception, log_error
+from namex.services.name_request.utils import handle_exception
 
 from namex.utils.logging import setup_logging
