+++ api/namex/services/virtual_word_condition/__init__.py
@@ -1,0 +1,1 @@
+from .virtual_word_condition import VirtualWordConditionService
